requirements.txt
camunda-external-task-client-python3==4.3.0
certifi==2021.10.8
chardet==3.0.4
idna==2.10
pydantic==1.9.0
requests==2.24.0
robotframework==5.0.1
typing_extensions==4.2.0
urllib3==1.25.11
